Update node_modules only if a file has changed
We'll use a MD5 hash in node_modules.hash to determine if the
node_modules folder needs to be extracted. A new hash is only
created when a file in the node_modules folder has changed.
The node_modules.hash file is created by the Node.js UWP project
system for Visual Studio.